requirements.txt
Flask
tensorflow==2.15.0
ultralytics
gunicorn
python-dotenv